+++ src/nanasqlite/__init__.py
@@ -22,6 +22,7 @@ Async Example:
 """
 
 from .async_core import AsyncNanaSQLite
+from .cache import CacheType
 from .core import NanaSQLite
 from .exceptions import (
     NanaSQLiteCacheError,
@@ -34,11 +35,12 @@ from .exceptions import (
     NanaSQLiteValidationError,
 )
 
-__version__ = "1.2.2"
+__version__ = "1.3.0dev0"
 __author__ = "Disnana"
 __all__ = [
     "NanaSQLite",
     "AsyncNanaSQLite",
+    "CacheType",
     "NanaSQLiteError",
     "NanaSQLiteValidationError",
     "NanaSQLiteDatabaseError",